Add GunFragment to control gun
Add fragment and layout for new Gun Fragment, which controls different features of the gun. Also added Adafruit's Bluetooth LE UART library to connect to the Arduino's BT adapter, and configured the library to connect to our BT adapter as well as added callbacks to the GunFragment to handle events.